Update code file structure
